Turn logging off by default
* Logging will now be turned off by default. This change is reflected in
  the readme as well.
* Remove unused code

+++ events/guildleave.js
@@ -6,6 +6,8 @@ module.exports = {
     name: 'guildDelete',
     execute(guild) {
 
+        if(!guild.name) return;
+
         console.log('\x1b[1m%s\x1b[0m', `Left guild: ${guild.name}`);
 
         Server.findOneAndRemove({
